Remove old HomeViewController
